osd: mark ourselves down if we are slow processing maps
If we get are not able to get the pgs to keep up with new osdmaps, ask the
mon to mark us down.  This will let us avoid all of the other work so that
we can catch up.  Also, once we are down we will drain peering_wq on each
map message.

Introduce a new tunable to control how far behind we have to be before this
kicks in.

Fixes: #7576
Signed-off-by: Sage Weil <sage@inktank.com>